pyproject.toml
[tool.poetry]
name = "redstar"
version = "0.1.0"
description = ""
authors = ["wangyuxin <wangyuxin@mokahr.com>"]
license = "MIT"
readme = "README.md"

[tool.ruff]
line-length = 128
select = [
    'E',
    'F',
    'I',
    'N'
]

[tool.poetry.dependencies]
python = "^3.10"
lmclient-core = "^0.3.0"
datasets = "^2.13.1"
rich = "^13.4.2"
scikit-learn = "^1.3.0"
typer = {extras = ["all"], version = "^0.9.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
blue = "^0.9.1"
ruff = "^0.0.265"
isort = "^5.12.0"
pyright = "^1.1.314"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
